attempt to work around openshift

+++ bot.py
@@ -270,7 +270,11 @@ def kill_if_disconnected(bot, timeout):
         time.sleep(timeout)
         if bot.idle_time() > timeout:
             bot.log("Idle for more than %ds. Exiting..." % timeout)
-            os.kill(os.getpid(), signal.SIGTERM)
+            if 'OPENSHIFT_APP_NAME' in os.environ:
+                import subprocess
+                subprocess.call(["ctl_all", "restart"])
+            else:
+                os.kill(os.getpid(), signal.SIGTERM)
             break
